+++ lib/lamby.rb
@@ -4,6 +4,7 @@ require 'base64'
 require 'active_support/all'
 require 'lamby/version'
 require 'lamby/sam_helpers'
+require 'lamby/rack'
 require 'lamby/rack_alb'
 require 'lamby/rack_api'
 require 'lamby/debug'
